Update to v2.05
Update to v2.05
install is not compatible with SupremeTeam Builds

+++ Quick_BGM_Install.sh
@@ -9,6 +9,7 @@ INSTALL_DIR=$(dirname "${SCRIPT_LOC}")
 MUSIC_DIR="$HOME/RetroPie/roms/music"
 MUSIC_DIR="${MUSIC_DIR/#~/$HOME}"
 MENU_DIR="$HOME/RetroPie/retropiemenu"
+STMENU_DIR="$HOME/RetroPie/retropiemenu/audiotools"
 AUTOSTART="/opt/retropie/configs/all/autostart.sh"
 PYGAME_PKG="python3-pygame"
 PSUTIL_PKG="omxplayer python-pygame mpg123 imagemagick python-urllib3 libpng12-0 fbi python-pip python3-pip python3-psutil"
@@ -34,15 +35,13 @@ infobox="${infobox}\n"
 infobox="${infobox}Overlay disappeared when you change resolutions? Set postion to Top-Left so you can see\n"
 infobox="${infobox}it then set it to desired postition, compatible with all resolutions.\n\n"
 infobox="${infobox}\n\n"
-
-dialog --backtitle "RetroPie Background Music Overlay Install Script v2.01" \
-	--title "RetroPie Background Music Overlay Install Script v2.01" \
+dialog --backtitle "RetroPie Background Music Overlay Install Script v2.05" \
+	--title "RetroPie Background Music Overlay Install Script v2.05" \
 	--msgbox "${infobox}" 35 110
-
 main_menu() {
     local choice
     while true; do
-        choice=$(dialog --colors --backtitle "RetroPie Background Music Overlay Install Script v2.01" --title " MAIN MENU " \
+        choice=$(dialog --colors --backtitle "RetroPie Background Music Overlay Install Script v2.05" --title " MAIN MENU " \
             --ok-label OK --cancel-label Exit \
             --menu "Choose An Option Below" 25 85 20 \
             01 "Minimal Install RPBGM Overlay No Music" \
@@ -59,6 +58,7 @@ main_menu() {
 }
 install_bgm() {
 minimum=1
+ST_STATUS=0
 clear
 prep_work
 setup
@@ -69,9 +69,7 @@ install_bgm_1() {
 minimum=0
 clear
 prep_work
-if [ -f "/home/pi/RetroPie/roms/music/arcade/arcade81.mp3" ]; then
-	echo "Found Music!"
-else
+if [ -f "/home/pi/RetroPie/roms/music/arcade/arcade81.mp3" ]; then 	echo "Found Music!"; else
 	gdown https://drive.google.com/uc?id=1-GLqdCNpH0i3zKRAJDOWwxfaP2gVGaC4 -O $HOME/retropie_music_overlay/bgm.zip
 	unzip -foq $HOME/retropie_music_overlay/bgm.zip -d $HOME/RetroPie
 fi
@@ -83,15 +81,11 @@ install_bgm_2() {
 minimum=0
 clear
 prep_work
-if [ -f "/home/pi/RetroPie/roms/music/arcade/arcade81.mp3" ]; then
-	echo "Found Music!"
-else
+if [ -f "/home/pi/RetroPie/roms/music/arcade/arcade81.mp3" ]; then echo "Found Music!"; else
 	gdown https://drive.google.com/uc?id=1-GLqdCNpH0i3zKRAJDOWwxfaP2gVGaC4 -O $HOME/retropie_music_overlay/bgm.zip
 	unzip -foq $HOME/retropie_music_overlay/bgm.zip -d $HOME/RetroPie
 fi
-if [ -f "/home/pi/RetroPie/roms/music/custom/3 Inches Of Blood- Deadly Sinners.mp3" ]; then
-	echo "Found Music!"
-else
+if [ -f "/home/pi/RetroPie/roms/music/custom/3 Inches Of Blood- Deadly Sinners.mp3" ]; then echo "Found Music!"; else
 	gdown https://drive.google.com/uc?id=1-BHwb4oT6GiwpRv7l3VLHuJLsRxScGNV -O $HOME/retropie_music_overlay/custombgm.zip
 	unzip -foq $HOME/retropie_music_overlay/bgm.zip -d $HOME/RetroPie
 	unzip -foq $HOME/retropie_music_overlay/custombgm.zip -d $HOME/RetroPie
@@ -104,15 +98,8 @@ exit
 prep_work() {
 ##### Install needed packages
 sudo apt-get update -y
-if sudo apt-get --simulate install python-pygame
-then 
-	sudo apt-get install -y python-pygame # to control music
-else
-	echo "
-
-Unable to install python-pygame, please update your system (\"sudo apt-get upgrade && sudo apt-get update\") and then try running this script again!
-	
-	"
+if sudo apt-get --simulate install $PYGAME_PKG; then sudo apt-get install -y $PYGAME_PKG; else
+	echo "Unable to install python-pygame, please update your system (\"sudo apt-get upgrade && sudo apt-get update\") and then try running this script again!"
 	exit
 fi
 sudo apt-get install -y $PSUTIL_PKG # to generate overlays
@@ -123,11 +110,12 @@ if [ -a $HOME/scripts/bgm/start.sc ]; then
 	pkill -STOP mpg123
 	sudo rm $HOME/scripts/bgm/start.sc
 fi
-currentuser=$(whoami) # Check user and then stop the script if root
-if [[ $currentuser == "root" ]]; then
-	echo "DON'T RUN THIS SCRIPT AS ROOT! USE './BGM_Install.sh' !"
-	exit
+if [ -d "$HOME/retropie_music_overlay" ]; then #delete folder if it is there
+	echo "$HOME/retropie_music_overlay Exists. Now Removing ..."
+	sudo rm -r $HOME/retropie_music_overlay
 fi
+currentuser=$(whoami) # Check user and then stop the script if root
+if [[ $currentuser == "root" ]]; then echo "DON'T RUN THIS SCRIPT AS ROOT! USE './BGM_Install.sh' !"; exit; fi
 ##### Download the files needed and install the script + utilities
 cd $HOME/retropie_music_overlay
 if [[ $currentuser == "pi" ]]; then #Use pngview if using Raspberry Pi
@@ -143,22 +131,17 @@ fi
 cd $HOME/retropie_music_overlay/
 sudo chmod +x $HOME/retropie_music_overlay/BGM.py
 sudo chmod +x $HOME/retropie_music_overlay/exit-splash
+sudo chmod +x $HOME/retropie_music_overlay/autostart.sh
 sudo chown $currentuser:$currentuser $HOME/retropie_music_overlay/BGM.py
 sudo chown $currentuser:$currentuser $HOME/retropie_music_overlay/exit-splash
+sudo chown $currentuser:$currentuser $HOME/retropie_music_overlay/autostart.sh
 sudo chmod 0777 $HOME/retropie_music_overlay/BGM.py
 sudo chmod 0777 $HOME/retropie_music_overlay/exit-splash
-if [ ! -d  "$MUSIC_DIR" ];
-then
-	mkdir $MUSIC_DIR
-else
-	echo "$MUSIC_DIR Exists!"
-fi	
-if [ -f "$HOME/BGM.py" ]; then #Remove old version if it is there
-	rm -f $HOME/BGM.py
-elif [ -f "$MUSIC_DIR/BGM.py" ]; then #Remove old version if it is there
-	rm -f $MUSIC_DIR/BGM.py
-elif [ -f "$INSTALL_DIR/BGM.py" ] ; then #Remove old version if it is there
-	rm -f $INSTALL_DIR/BGM.py
+sudo chmod 0777 $HOME/retropie_music_overlay/autostart.sh
+if [ ! -d  "$MUSIC_DIR" ]; then mkdir $MUSIC_DIR; else echo "$MUSIC_DIR Exists!"; fi	
+if [ -f "$HOME/BGM.py" ]; then rm -f $HOME/BGM.py
+elif [ -f "$MUSIC_DIR/BGM.py" ]; then rm -f $MUSIC_DIR/BGM.py
+elif [ -f "$INSTALL_DIR/BGM.py" ] ; then rm -f $INSTALL_DIR/BGM.py
 fi
 if [ $INSTALL_DIR ]; then mkdir $INSTALL_DIR; fi
 cp BGM.py $INSTALL_DIR
@@ -170,45 +153,38 @@ sudo cp $HOME/retropie_music_overlay/Pixel.otf /usr/share/fonts/opentype/
 sudo cp $HOME/retropie_music_overlay/GROBOLD.ttf /usr/share/fonts/truetype/
 ##### Add menu options for BGM Overlay Controls
 cp $HOME/retropie_music_overlay/backgroundmusic.png $MENU_DIR/icons/
-if [ -f "$MENU_DIR/bgmcustomoptions.sh" ]; then # Remove old version if it is there
-	sudo rm -f $MENU_DIR/bgmcustomoptions.sh
-fi
+if [ -f "$MENU_DIR/bgmcustomoptions.sh" ]; then sudo rm -f $MENU_DIR/bgmcustomoptions.sh; fi
+if [ -f "$STMENU_DIR/bgmcustomoptions.sh" ]; then sudo rm -f $STMENU_DIR/bgmcustomoptions.sh; fi
+if [ -d "$STMENU_DIR" ]; then RP_MENU=$STMENU_DIR; else RP_MENU=$MENU_DIR; fi
 if [ "$minimum" = "1" ]; then
 	sudo chmod +x $HOME/retropie_music_overlay/bgmcustomoptions-minimum.sh
 	sudo chown $currentuser:$currentuser $HOME/retropie_music_overlay/bgmcustomoptions-minimum.sh
 	sudo chmod 0777 $HOME/retropie_music_overlay/bgmcustomoptions-minimum.sh
-	cp bgmcustomoptions-minimum.sh $MENU_DIR/bgmcustomoptions.sh
+	cp bgmcustomoptions-minimum.sh $RP_MENU/bgmcustomoptions.sh
 else
 	sudo chmod +x $HOME/retropie_music_overlay/bgmcustomoptions.sh
 	sudo chown $currentuser:$currentuser $HOME/retropie_music_overlay/bgmcustomoptions.sh
 	sudo chmod 0777 $HOME/retropie_music_overlay/bgmcustomoptions.sh
-	cp bgmcustomoptions.sh $MENU_DIR
+	cp bgmcustomoptions.sh $RP_MENU
 fi
-if [ ! -s $MENU_DIR/gamelist.xml ] # Remove gamelist.xml if file size is 0
-then
-	sudo rm -f $MENU_DIR/gamelist.xml
-fi
-if [ ! -f "$MENU_DIR/gamelist.xml" ]; # If file doesn't exist, copy gamelist.xml to folder
-then
-	cp /opt/retropie/configs/all/emulationstation/gamelists/retropie/gamelist.xml $MENU_DIR/gamelist.xml
-fi
+if [ ! -s $MENU_DIR/gamelist.xml ]; then sudo rm -f $MENU_DIR/gamelist.xml; fi
+if [ ! -f "$MENU_DIR/gamelist.xml" ]; then cp /opt/retropie/configs/all/emulationstation/gamelists/retropie/gamelist.xml $MENU_DIR/gamelist.xml; fi
+if [ -d "$STMENU_DIR" ]; then
+CONTENT1="<game>\n<path>./audiotools/bgmcustomoptions.sh</path>\n<name>BGM Overlay Options</name>\n<desc>Toggles background music options such as music ON/OFF and volume control. Set and play MP3/OGG files during menu navigation in both Emulation Station and Attract Mode. A Few special new folders have been created in the /music directory called "arcade", "bttf", "st", "uvf", "venom", and this last one "custom" is for placing your own MP3/OGG files into. Once you place your music files into this folder and enable it, the music will automatically begin playing.</desc>\n<image>./icons/backgroundmusic.png</image>\n<releasedate>20211110T000251</releasedate>\n<developer>Livewire, madmodder123, thepitster</developer>\n<publisher>Livewire, madmodder123, thepitster</publisher>\n<genre>BGM Control Script</genre>\n</game>"
+C1=$(echo $CONTENT1 | sed 's/\//\\\//g')
+else
 CONTENT1="<game>\n<path>./bgmcustomoptions.sh</path>\n<name>BGM Overlay Options</name>\n<desc>Toggles background music options such as music ON/OFF and volume control. Set and play MP3/OGG files during menu navigation in both Emulation Station and Attract Mode. A Few special new folders have been created in the /music directory called "arcade", "bttf", "st", "uvf", "venom", and this last one "custom" is for placing your own MP3/OGG files into. Once you place your music files into this folder and enable it, the music will automatically begin playing.</desc>\n<image>./icons/backgroundmusic.png</image>\n<releasedate>20211110T000251</releasedate>\n<developer>Livewire, madmodder123, thepitster</developer>\n<publisher>Livewire, madmodder123, thepitster</publisher>\n<genre>BGM Control Script</genre>\n</game>"
 C1=$(echo $CONTENT1 | sed 's/\//\\\//g')
-if grep -q bgmcustomoptions.sh "$MENU_DIR/gamelist.xml"; then # Check if menu entry is already there or not
-	echo "gamelist.xml entry confirmed"
+fi
+if grep -q bgmcustomoptions.sh "$MENU_DIR/gamelist.xml"; then echo "gamelist.xml entry confirmed"
 else
 	sed "/<\/gameList>/ s/.*/${C1}\n&/" $MENU_DIR/gamelist.xml > $HOME/temp
 	cat $HOME/temp > $MENU_DIR/gamelist.xml
 	rm -f $HOME/temp
 fi
-if ! grep -q '#(nohup python /home/pi/.rpbgmo/BGM.py & > /dev/null 2>&1)' "$AUTOSTART"; then
-	sed -i 's/\#(nohup python/(nohup python/g' $AUTOSTART
-fi
-if ! grep -q '(nohup python /home/pi/.rpbgmo/BGM.py > /dev/null 2>&1) &' "$AUTOSTART"; then
-	sed -i -E '$ i\(nohup python /home/pi/.rpbgmo/BGM.py > /dev/null 2>&1) &' $AUTOSTART
-else
-	echo "BGM already running at boot!"
-fi
+if ! grep -q '#(nohup python /home/pi/.rpbgmo/BGM.py & > /dev/null 2>&1)' "$AUTOSTART"; then sed -i 's/\#(nohup python/(nohup python/g' $AUTOSTART; fi
+if ! grep -q '(nohup python /home/pi/.rpbgmo/BGM.py > /dev/null 2>&1) &' "$AUTOSTART"; then sed -i -E '$ i\(nohup python /home/pi/.rpbgmo/BGM.py > /dev/null 2>&1) &' $AUTOSTART
+else echo "BGM already running at boot!"; fi
 ##### Setting up Splash & Exit Screens
 cp "$HOME/retropie_music_overlay/BGM Folder Diabled.mp3"  $INSTALL_DIR
 cp $HOME/retropie_music_overlay/splashscreens/JarvisSplash.mp4 $HOME/RetroPie/splashscreens/
@@ -218,17 +194,16 @@ mkdir -p /opt/retropie/configs/all/emulationstation/scripts/reboot
 mkdir -p /opt/retropie/configs/all/emulationstation/scripts/shutdown
 cp $HOME/retropie_music_overlay/exit-splash /opt/retropie/configs/all/emulationstation/scripts/reboot/
 cp $HOME/retropie_music_overlay/exit-splash /opt/retropie/configs/all/emulationstation/scripts/shutdown/
-if ! grep -q '/opt/retropie/configs/all/emulationstation/scripts/shutdown/exit-splash' "$AUTOSTART"; then
+mv -f /opt/retropie/configs/all/autostart.sh /opt/retropie/configs/all/autostart.sh.BACKUP
+cp $HOME/retropie_music_overlay/autostart.sh /opt/retropie/configs/all/
 	sed -i -E '$a\/opt/retropie/configs/all/emulationstation/scripts/shutdown/exit-splash' $AUTOSTART
-else
-	echo "Exit Splash Already Set!"
-fi
+else echo "Exit Splash Already Set!"; fi
 cd $HOME
 ##### Explain stuff to the user
 printf "\n\n\n"
 echo "Place your personal music files in $MUSIC_DIR/custom/"
-echo "Run $MENU_DIR/bgmcustomoptions.sh or navigate to:"
-echo "Retropie > Background Music Options, for more options!"
+echo "Run $RP_MENU/bgmcustomoptions.sh or navigate to:"
+if [ -d "$STMENU_DIR" ]; then echo "Retropie > Audiotools > Background Music Options, for more options!";else echo "Retropie > Background Music Options, for more options!"; fi
 echo "BGM has been set up to run automatically when the device boots!
 
 "
